+++ test/amplitude.js
@@ -12,7 +12,7 @@ describe('Amplitude', function() {
   var Identify = require('../src/identify.js');
   var Revenue = require('../src/revenue.js');
   var apiKey = '000000';
-  var keySuffix = '_' + apiKey.slice(0,6);
+  var keySuffix = '_' + apiKey;
   var userId = 'user';
   var amplitude;
   var server;
@@ -34,9 +34,13 @@ describe('Amplitude', function() {
     localStorage.clear();
     sessionStorage.clear();
     cookie.remove(amplitude.options.cookieName);
-    cookie.remove(amplitude.options.cookieName + keySuffix);
     cookie.remove(amplitude.options.cookieName + '_app1');
     cookie.remove(amplitude.options.cookieName + '_app2');
+    cookie.remove(amplitude.options.cookieName + keySuffix);
+    cookie.remove(amplitude.options.cookieName + keySuffix + '_app1');
+    cookie.remove(amplitude.options.cookieName + keySuffix + '_app2');
+    cookie.remove(amplitude.options.cookieName + '_1_app1');
+    cookie.remove(amplitude.options.cookieName + '_2_app2');
     cookie.reset();
   }
 
@@ -134,20 +138,21 @@ describe('Amplitude', function() {
 
       // verify separate localstorages
       assert.deepEqual(
-        JSON.parse(localStorage.getItem('amplitude_unsent'))[0].event_type, 'amplitude event'
+        JSON.parse(localStorage.getItem('amplitude_unsent_' + apiKey))[0].event_type, 'amplitude event'
       );
       assert.deepEqual(
-        JSON.parse(localStorage.getItem('amplitude_unsent'))[1].event_type, 'amplitude event2'
+        JSON.parse(localStorage.getItem('amplitude_unsent_' + apiKey))[1].event_type, 'amplitude event2'
       );
-      assert.equal(localStorage.getItem('amplitude_unsent_identify'), JSON.stringify([]));
-      assert.equal(localStorage.getItem('amplitude_unsent_app1'), JSON.stringify([]));
+      assert.equal(localStorage.getItem('amplitude_unsent_identify_' + apiKey), JSON.stringify([]));
+      assert.equal(localStorage.getItem('amplitude_unsent_1_app1'), JSON.stringify([]));
       assert.deepEqual(
-        JSON.parse(localStorage.getItem('amplitude_unsent_identify_app1'))[0].user_properties, {'$set':{'key':'value'}}
+        JSON.parse(localStorage.getItem('amplitude_unsent_identify_1_app1'))[0].user_properties,
+        {'$set':{'key':'value'}}
       );
       assert.equal(
-        JSON.parse(localStorage.getItem('amplitude_unsent_app2'))[0].event_type, 'app2 event'
+        JSON.parse(localStorage.getItem('amplitude_unsent_2_app2'))[0].event_type, 'app2 event'
       );
-      assert.equal(localStorage.getItem('amplitude_unsent_identify_app2'), JSON.stringify([]));
+      assert.equal(localStorage.getItem('amplitude_unsent_identify_2_app2'), JSON.stringify([]));
 
       // verify separate apiKeys in server requests
       assert.lengthOf(server.requests, 3);
@@ -155,13 +160,13 @@ describe('Amplitude', function() {
       assert.equal(JSON.parse(querystring.parse(server.requests[2].requestBody).client), 2);
 
       // verify separate cookie data
-      var cookieData = cookie.get(amplitude.options.cookieName);
+      var cookieData = cookie.get(amplitude.options.cookieName + keySuffix);
       assert.equal(cookieData.deviceId, amplitude.options.deviceId);
 
-      var cookieData1 = cookie.get(app1.options.cookieName + '_app1');
+      var cookieData1 = cookie.get(app1.options.cookieName + '_1_app1');
       assert.equal(cookieData1.deviceId, app1.options.deviceId);
 
-      var cookieData2 = cookie.get(app2.options.cookieName + '_app2');
+      var cookieData2 = cookie.get(app2.options.cookieName + '_2_app2');
       assert.equal(cookieData2.deviceId, app2.options.deviceId);
     });
 
@@ -178,7 +183,7 @@ describe('Amplitude', function() {
         identifyId: 60,
         sequenceNumber: 70
       }
-      cookie.set(amplitude.options.cookieName, cookieData);
+      cookie.set(amplitude.options.cookieName + keySuffix, cookieData);
 
       // default instance loads from existing cookie
       var app = amplitude.getInstance();
@@ -197,7 +202,6 @@ describe('Amplitude', function() {
       assert.notEqual(app1.options.deviceId, 'test_device_id');
       assert.isNull(app1.options.userId);
       assert.isFalse(app1.options.optOut);
-      console.log(app1._sessionId);
       assert.isTrue(app1._sessionId >= now);
       assert.isTrue(app1._lastEventTime >= now);
       assert.equal(app1._eventId, 0);
@@ -274,7 +278,7 @@ describe('Amplitude', function() {
 
     it('should set cookie', function() {
       amplitude.init(apiKey, userId);
-      var stored = cookie.get(amplitude.options.cookieName);
+      var stored = cookie.get(amplitude.options.cookieName + keySuffix);
       assert.property(stored, 'deviceId');
       assert.propertyVal(stored, 'userId', userId);
       assert.lengthOf(stored.deviceId, 37); // increase deviceId length by 1 for 'R' character
@@ -309,6 +313,7 @@ describe('Amplitude', function() {
       var userId = 'test_user_id';
 
       assert.isNull(cookie.get(amplitude.options.cookieName));
+      assert.isNull(cookie.get(amplitude.options.cookieName + keySuffix));
       localStorage.setItem('amplitude_deviceId' + keySuffix, deviceId);
       localStorage.setItem('amplitude_userId' + keySuffix, userId);
       localStorage.setItem('amplitude_optOut' + keySuffix, true);
@@ -318,7 +323,7 @@ describe('Amplitude', function() {
       assert.equal(amplitude.options.userId, userId);
       assert.isTrue(amplitude.options.optOut);
 
-      var cookieData = cookie.get(amplitude.options.cookieName);
+      var cookieData = cookie.get(amplitude.options.cookieName + keySuffix);
       assert.equal(cookieData.deviceId, deviceId);
       assert.equal(cookieData.userId, userId);
       assert.isTrue(cookieData.optOut);
@@ -328,6 +333,7 @@ describe('Amplitude', function() {
       var now = new Date().getTime();
 
       assert.isNull(cookie.get(amplitude.options.cookieName));
+      assert.isNull(cookie.get(amplitude.options.cookieName + keySuffix));
       localStorage.setItem('amplitude_sessionId', now);
       localStorage.setItem('amplitude_lastEventTime', now);
       localStorage.setItem('amplitude_lastEventId', 3000);
@@ -342,7 +348,7 @@ describe('Amplitude', function() {
       assert.equal(amplitude.getInstance()._identifyId, 4000);
       assert.equal(amplitude.getInstance()._sequenceNumber, 5000);
 
-      var cookieData = cookie.get(amplitude.options.cookieName);
+      var cookieData = cookie.get(amplitude.options.cookieName + keySuffix);
       assert.equal(cookieData.sessionId, now);
       assert.equal(cookieData.lastEventTime, amplitude.getInstance()._lastEventTime);
       assert.equal(cookieData.eventId, 3000);
@@ -382,6 +388,35 @@ describe('Amplitude', function() {
       assert.equal(amplitude.getInstance()._eventId, 50);
       assert.equal(amplitude.getInstance()._identifyId, 60);
       assert.equal(amplitude.getInstance()._sequenceNumber, 40);
+    });
+
+    it('should migrate cookie data from old scope to new scope', function(){
+      var now = new Date().getTime();
+      assert.isNull(cookie.get(amplitude.options.cookieName));
+      assert.isNull(cookie.get(amplitude.options.cookieName + keySuffix));
+
+      var cookieData = {
+        deviceId: 'test_device_id',
+        userId: 'test_user_id4',
+        optOut: false,
+        sessionId: now,
+        lastEventTime: now,
+        eventId: 50,
+        identifyId: 60,
+        sequenceNumber: 15,
+      }
+
+      cookie.set(amplitude.options.cookieName, cookieData);
+      amplitude.init(apiKey);
+      var newCookieData = cookie.get(amplitude.options.cookieName + keySuffix);
+      assert.equal(cookieData.deviceId, newCookieData.deviceId);
+      assert.equal(cookieData.userId, newCookieData.userId);
+      assert.isFalse(newCookieData.optOut);
+      assert.equal(cookieData.sessionId, newCookieData.sessionId);
+      assert.isTrue(newCookieData.lastEventTime >= cookieData.lastEventTime);
+      assert.equal(cookieData.eventId, newCookieData.eventId);
+      assert.equal(cookieData.identifyId, newCookieData.identifyId);
+      assert.equal(cookieData.sequenceNumber, newCookieData.sequenceNumber);
     });
 
     it('should skip the migration if the new cookie already has deviceId, sessionId, lastEventTime', function() {
@@ -443,7 +478,7 @@ describe('Amplitude', function() {
         identifyId: 60,
         sequenceNumber: 70
       }
-      cookie.set(amplitude2.options.cookieName, cookieData);
+      cookie.set(amplitude2.options.cookieName + keySuffix, cookieData);
 
       clock.tick(10);
       amplitude2.init(apiKey);
@@ -505,8 +540,8 @@ describe('Amplitude', function() {
         '"event_properties":{},"user_properties":{"$set":{"age":30,"city":"San Francisco, CA"}},"uuid":"' +
         'c50e1be4-7976-436a-aa25-d9ee38951082","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number"' +
         ':131, "groups":{}}]';
-      localStorage.setItem('amplitude_unsent', existingEvent);
-      localStorage.setItem('amplitude_unsent_identify', existingIdentify);
+      localStorage.setItem('amplitude_unsent_' + apiKey, existingEvent);
+      localStorage.setItem('amplitude_unsent_identify_' + apiKey, existingIdentify);
 
       var amplitude2 = new Amplitude();
       amplitude2.init(apiKey, null, {batchEvents: true});
@@ -516,8 +551,8 @@ describe('Amplitude', function() {
       assert.deepEqual(amplitude2.getInstance()._unsentIdentifys, JSON.parse(existingIdentify));
 
       // check local storage keys are still same for default instance
-      assert.equal(localStorage.getItem('amplitude_unsent'), existingEvent);
-      assert.equal(localStorage.getItem('amplitude_unsent_identify'), existingIdentify);
+      assert.equal(localStorage.getItem('amplitude_unsent_' + apiKey), existingEvent);
+      assert.equal(localStorage.getItem('amplitude_unsent_identify_' + apiKey), existingIdentify);
     });
 
     it('should validate event properties when loading saved events from localStorage', function() {
@@ -533,7 +568,7 @@ describe('Amplitude', function() {
         '[0,1,2,"3"],"nested_array":["a",{"key":"value"},["b"]],"object":{"key":"value"},"nested_object":' +
         '{"k":"v","l":[0,1],"o":{"k2":"v2","l2":["e2",{"k3":"v3"}]}}},"user_properties":{},"uuid":"650407a1-d705-' +
         '47a0-8918-b4530ce51f89","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number":5}]'
-      localStorage.setItem('amplitude_unsent', existingEvents);
+      localStorage.setItem('amplitude_unsent_' + apiKey, existingEvents);
 
       var amplitude2 = new Amplitude();
       amplitude2.init(apiKey, null, {batchEvents: true});
@@ -562,7 +597,7 @@ describe('Amplitude', function() {
         '[0,1,2,"3"],"nested_array":["a",{"key":"value"},["b"]],"object":{"key":"value"},"nested_object":' +
         '{"k":"v","l":[0,1],"o":{"k2":"v2","l2":["e2",{"k3":"v3"}]}}}},"event_properties":{},"uuid":"650407a1-d705-' +
         '47a0-8918-b4530ce51f89","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number":5}]'
-      localStorage.setItem('amplitude_unsent_identify', existingEvents);
+      localStorage.setItem('amplitude_unsent_identify_' + apiKey, existingEvents);
 
       var amplitude2 = new Amplitude();
       amplitude2.init(apiKey, null, {batchEvents: true});
@@ -593,8 +628,8 @@ describe('Amplitude', function() {
         '"event_properties":{},"user_properties":{"$set":{"age":30,"city":"San Francisco, CA"}},"uuid":"' +
         'c50e1be4-7976-436a-aa25-d9ee38951082","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number"' +
         ':131}]';
-      localStorage.setItem('amplitude_unsent', existingEvent);
-      localStorage.setItem('amplitude_unsent_identify', existingIdentify);
+      localStorage.setItem('amplitude_unsent_' + apiKey, existingEvent);
+      localStorage.setItem('amplitude_unsent_identify_' + apiKey, existingIdentify);
 
       var amplitude2 = new Amplitude();
       amplitude2.init(apiKey, null, {batchEvents: true, eventUploadThreshold: 2});
@@ -606,8 +641,8 @@ describe('Amplitude', function() {
       assert.deepEqual(amplitude2.getInstance()._unsentIdentifys, []);
 
       // check local storage keys are still same
-      assert.equal(localStorage.getItem('amplitude_unsent'), JSON.stringify([]));
-      assert.equal(localStorage.getItem('amplitude_unsent_identify'), JSON.stringify([]));
+      assert.equal(localStorage.getItem('amplitude_unsent_' + apiKey), JSON.stringify([]));
+      assert.equal(localStorage.getItem('amplitude_unsent_identify_' + apiKey), JSON.stringify([]));
 
       // check request
       assert.lengthOf(server.requests, 1);
@@ -619,18 +654,18 @@ describe('Amplitude', function() {
 
     it('should validate event properties when loading saved events from localStorage', function() {
       var existingEvents = '[{"device_id":"15a82aaa-0d9e-4083-a32d-2352191877e6","user_id":"15a82aaa-0d9e-4083-a32d' +
-          '-2352191877e6","timestamp":1455744744413,"event_id":2,"session_id":1455744733865,"event_type":"clicked",' +
-          '"version_name":"Web","platform":"Web","os_name":"Chrome","os_version":"48","device_model":"Mac","language"' +
-          ':"en-US","api_properties":{},"event_properties":"{}","user_properties":{},"uuid":"1b8859d9-e91e-403e-92d4-' +
-          'c600dfb83432","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number":4},{"device_id":"15a82a' +
-          'aa-0d9e-4083-a32d-2352191877e6","user_id":"15a82aaa-0d9e-4083-a32d-2352191877e6","timestamp":1455744746295,' +
-          '"event_id":3,"session_id":1455744733865,"event_type":"clicked","version_name":"Web","platform":"Web",' +
-          '"os_name":"Chrome","os_version":"48","device_model":"Mac","language":"en-US","api_properties":{},' +
-          '"event_properties":{"10":"false","bool":true,"null":null,"string":"test","array":' +
-          '[0,1,2,"3"],"nested_array":["a",{"key":"value"},["b"]],"object":{"key":"value"},"nested_object":' +
-          '{"k":"v","l":[0,1],"o":{"k2":"v2","l2":["e2",{"k3":"v3"}]}}},"user_properties":{},"uuid":"650407a1-d705-' +
-          '47a0-8918-b4530ce51f89","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number":5}]';
-      localStorage.setItem('amplitude_unsent', existingEvents);
+        '-2352191877e6","timestamp":1455744744413,"event_id":2,"session_id":1455744733865,"event_type":"clicked",' +
+        '"version_name":"Web","platform":"Web","os_name":"Chrome","os_version":"48","device_model":"Mac","language"' +
+        ':"en-US","api_properties":{},"event_properties":"{}","user_properties":{},"uuid":"1b8859d9-e91e-403e-92d4-' +
+        'c600dfb83432","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number":4},{"device_id":"15a82a' +
+        'aa-0d9e-4083-a32d-2352191877e6","user_id":"15a82aaa-0d9e-4083-a32d-2352191877e6","timestamp":1455744746295,' +
+        '"event_id":3,"session_id":1455744733865,"event_type":"clicked","version_name":"Web","platform":"Web",' +
+        '"os_name":"Chrome","os_version":"48","device_model":"Mac","language":"en-US","api_properties":{},' +
+        '"event_properties":{"10":"false","bool":true,"null":null,"string":"test","array":' +
+        '[0,1,2,"3"],"nested_array":["a",{"key":"value"},["b"]],"object":{"key":"value"},"nested_object":' +
+        '{"k":"v","l":[0,1],"o":{"k2":"v2","l2":["e2",{"k3":"v3"}]}}},"user_properties":{},"uuid":"650407a1-d705-' +
+        '47a0-8918-b4530ce51f89","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number":5}]';
+      localStorage.setItem('amplitude_unsent_' + apiKey, existingEvents);
 
       var amplitude2 = new Amplitude();
       amplitude2.init(apiKey, null, {
@@ -659,6 +694,61 @@ describe('Amplitude', function() {
       // check that event loaded into memory
       assert.deepEqual(amplitude2.getInstance()._unsentEvents[0].event_properties, {});
       assert.deepEqual(amplitude2.getInstance()._unsentEvents[1].event_properties, expected);
+    });
+
+    it('should migrate saved events from old localstorage keys to new scoped keys', function() {
+      var existingEvent = '[{"device_id":"test_device_id","user_id":"test_user_id","timestamp":1453769146589,' +
+        '"event_id":49,"session_id":1453763315544,"event_type":"clicked","version_name":"Web","platform":"Web"' +
+        ',"os_name":"Chrome","os_version":"47","device_model":"Mac","language":"en-US","api_properties":{},' +
+        '"event_properties":{},"user_properties":{},"uuid":"3c508faa-a5c9-45fa-9da7-9f4f3b992fb0","library"' +
+        ':{"name":"amplitude-js","version":"2.9.0"},"sequence_number":130,"groups":{}}]';
+      var existingIdentify = '[{"device_id":"test_device_id","user_id":"test_user_id","timestamp":1453769338995,' +
+        '"event_id":82,"session_id":1453763315544,"event_type":"$identify","version_name":"Web","platform":"Web"' +
+        ',"os_name":"Chrome","os_version":"47","device_model":"Mac","language":"en-US","api_properties":{},' +
+        '"event_properties":{},"user_properties":{"$set":{"age":30,"city":"San Francisco, CA"}},"uuid":"' +
+        'c50e1be4-7976-436a-aa25-d9ee38951082","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number"' +
+        ':131,"groups":{}}]';
+
+      localStorage.setItem('amplitude_unsent', existingIdentify);
+      localStorage.setItem('amplitude_unsent_' + apiKey, existingEvent);
+
+      var amplitude2 = new Amplitude();
+      amplitude2.init(apiKey, null, {
+        batchEvents: true
+      });
+
+      // verify that events loaded in proper order and that keys were migrated
+      assert.lengthOf(amplitude2.getInstance()._unsentEvents, 2);
+      assert.equal(amplitude2.getInstance()._unsentEvents[0]['event_type'], '$identify');  // loads old one first
+      assert.equal(amplitude2.getInstance()._unsentEvents[1]['event_type'], 'clicked');
+      assert.equal(localStorage.getItem('amplitude_unsent'), undefined);  // old key is deleted
+      assert.deepEqual(
+        JSON.parse(localStorage.getItem('amplitude_unsent_' + apiKey)), amplitude2.getInstance()._unsentEvents
+      );
+    });
+
+    it('should migrate saved identifys from old localstorage keys to new scoped keys', function() {
+      var existingIdentify = '[{"device_id":"test_device_id","user_id":"test_user_id","timestamp":1453769338995,' +
+        '"event_id":82,"session_id":1453763315544,"event_type":"$identify","version_name":"Web","platform":"Web"' +
+        ',"os_name":"Chrome","os_version":"47","device_model":"Mac","language":"en-US","api_properties":{},' +
+        '"event_properties":{},"user_properties":{"$set":{"age":30,"city":"San Francisco, CA"}},"uuid":"' +
+        'c50e1be4-7976-436a-aa25-d9ee38951082","library":{"name":"amplitude-js","version":"2.9.0"},"sequence_number"' +
+        ':131,"groups":{}}]';
+
+      localStorage.setItem('amplitude_unsent_identify', existingIdentify);
+
+      var amplitude2 = new Amplitude();
+      amplitude2.init(apiKey, null, {
+        batchEvents: true
+      });
+
+      assert.lengthOf(amplitude2.getInstance()._unsentIdentifys, 1);
+      assert.equal(amplitude2.getInstance()._unsentIdentifys[0]['event_type'], '$identify');
+      assert.equal(localStorage.getItem('amplitude_unsent_identify'), undefined);  // old key is deleted
+      assert.deepEqual(
+        JSON.parse(localStorage.getItem('amplitude_unsent_identify_' + apiKey)),
+        amplitude2.getInstance()._unsentIdentifys
+      );
     });
   });
 
@@ -873,7 +963,7 @@ describe('setVersionName', function() {
     it('should store device id in cookie', function() {
       amplitude.init(apiKey, null, {'deviceId': 'fakeDeviceId'});
       amplitude.setDeviceId('deviceId');
-      var stored = cookie.get(amplitude.options.cookieName);
+      var stored = cookie.get(amplitude.options.cookieName + keySuffix);
       assert.propertyVal(stored, 'deviceId', 'deviceId');
     });
   });
@@ -1156,7 +1246,7 @@ describe('setVersionName', function() {
     });
 
     it('should save events', function() {
-      amplitude.init(apiKey, null, {saveEvents: true});
+      amplitude.init(apiKey, null, {saveEvents: true, batchEvents: true});
       amplitude.logEvent('Event', {index: 1});
       amplitude.logEvent('Event', {index: 2});
       amplitude.logEvent('Event', {index: 3});
@@ -1866,7 +1956,7 @@ describe('setVersionName', function() {
       clock.tick(20);
       amplitude2.setUserProperties({'key':'value'}); // identify event at time 30
 
-      var cookieData = JSON.parse(localStorage.getItem('amp_cookiestore_amplitude_id'));
+      var cookieData = JSON.parse(localStorage.getItem('amp_cookiestore_amplitude_id_000000'));
       assert.deepEqual(cookieData, {
         'deviceId': deviceId,
         'userId': null,